exchange_like/urls.py
"""django_celery URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/1.9/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.conf.urls import url, include
    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
"""
from django.conf.urls import include, url
from django.contrib import admin
from django.conf import settings
from django.conf.urls.static import static
from django.views.decorators.cache import cache_page
from adminplus.sites import AdminSitePlus

import exchange.urls

urlpatterns = [
    url(r'^admin/', include(admin.site.urls)),
    url(r'^', include(exchange.urls, namespace='exchange_like')),
   # url(r'^exchange_like/', include(exchange_like.urls, namespace='exchange_like')),
    #url(r'^schedule_task/', include(schedule_task.urls, namespace='schedule_task')),
]

# User-uploaded files like profile pics need to be served in development
urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
