requirements.txt
Django>=1.9
django_celery
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
django-bootstrap3>=7.0
django-authtools>=1.4.0
easy-thumbnails>=2.3
requests>=2.9
django-mobile>=0.6
django-cors-headers>=1.1.0
fb>=0.3.0
django-adminplus>=0.5
python-oembed>=0.2.4
django-redis-cache>=1.6.5
django-chartjs>=1.0
mysqlclient
